# Early exit for empty HashMap (issue #38880)