@bluss 2018-02-09
Nice work. Especially that your rearrangement shows that this is a win we don't really have to pay anything for. My suggestions are in the previous comment.