@bluss 2018-02-12
I think there is a refactoring we can do to get the full `.size() != 0` benefit, except when using `.entry()`.

1. in `search(&self, ..)` we are never inserting, so we could use the size check?
2. Lots of `search_mut` users check the `size()` up front already (see below); it could be refactored so that all possible `search_mut` users have the size check (and entry uses the one with just capacity check).

https://github.com/rust-lang/rust/blob/fd78621717e4f9f73b41256627bfe3a83aa5e660/src/libstd/collections/hash/map.rs#L2658-L2664