@technicalguy 2018-02-13
I'm not sure I'm entirely happy with this...I'd like to enforce the check through the type system if possible...