@technicalguy 2018-02-13
@bluss good observation. The `entry` method is the only one which requires an `InternalEntry` return type. And in fact it originally never used the `search[_mut]` method. All the other users of `search[_mut]` call `.into_occupied_bucket()`, so they don't require an internal entry, and thus the `search[_mut]` methods can be updated to return an `Option<FullBucket<...>>`. This means that the `search[_mut]` methods can be updated to check if the map is empty (contains no elements), not just that the capacity is zero (is unallocated).

Furthermore, this means that there is hope for a linear search optimisation, since we don't actually need to return an `InternalEntry` in most cases.