@bluss 2018-02-13
I think there are a few more things that can be streamlined if one goes over all uses of search_mut, search_hashed etc. For example `search_hashed` checks the capacity == 0 case but its both two callers are from a place where capacity > 0. Removing those checks can be weighed in a different PR — plainly removing them would make the code less defensive w.r.t memory safety bugs.